+++ slice/doc.go
@@ -71,6 +71,9 @@ func _() {
 	_ = Chunk[int]
 	_ = Compact[int]
 	_ = Contains[int]
+	_ = FromMap[int, int]
+	_ = FromSet[int]
+	_ = GroupBy[int, int]
 	_ = SortBy[int, int]
 	_ = SortByDesc[int, int]